# Conflicts:
#	backend/src/main/java/io/f1/backend/domain/game/app/RoomService.java
#	backend/src/main/java/io/f1/backend/domain/game/websocket/GameSocketController.java
#	backend/src/main/java/io/f1/backend/global/exception/errorcode/RoomErrorCode.java
#	backend/src/test/java/io/f1/backend/domain/game/app/RoomServiceTests.java

+++ backend/src/main/java/io/f1/backend/domain/question/app/QuestionService.java
@@ -10,12 +10,17 @@ import io.f1.backend.domain.question.entity.Question;
 import io.f1.backend.domain.question.entity.TextQuestion;
 import io.f1.backend.domain.quiz.entity.Quiz;
 import io.f1.backend.global.exception.CustomException;
+import io.f1.backend.global.exception.errorcode.AuthErrorCode;
 import io.f1.backend.global.exception.errorcode.QuestionErrorCode;
+import io.f1.backend.global.security.enums.Role;
+import io.f1.backend.global.util.SecurityUtils;
 
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -47,8 +52,19 @@ public class QuestionService {
                         .orElseThrow(
                                 () -> new CustomException(QuestionErrorCode.QUESTION_NOT_FOUND));
 
+        verifyUserAuthority(question.getQuiz());
+
         TextQuestion textQuestion = question.getTextQuestion();
         textQuestion.changeContent(content);
+    }
+
+    private static void verifyUserAuthority(Quiz quiz) {
+        if (SecurityUtils.getCurrentUserRole() == Role.ADMIN) {
+            return;
+        }
+        if (!Objects.equals(SecurityUtils.getCurrentUserId(), quiz.getCreator().getId())) {
+            throw new CustomException(AuthErrorCode.FORBIDDEN);
+        }
     }
 
     @Transactional
@@ -62,6 +78,8 @@ public class QuestionService {
                         .orElseThrow(
                                 () -> new CustomException(QuestionErrorCode.QUESTION_NOT_FOUND));
 
+        verifyUserAuthority(question.getQuiz());
+
         question.changeAnswer(answer);
     }
 
@@ -73,6 +91,8 @@ public class QuestionService {
                         .findById(questionId)
                         .orElseThrow(
                                 () -> new CustomException(QuestionErrorCode.QUESTION_NOT_FOUND));
+
+        verifyUserAuthority(question.getQuiz());
 
         questionRepository.delete(question);
     }
